ShopSearch Service단 수정 및 기능 추가
